[Kernel] Delete kernel-defaults JsonUtils; move rowFromJson to DefaultJsonRow
Follow-up to fully remove the kernel-defaults JsonUtils duplication instead
of leaving a delegating wrapper. Deletes
kernel-defaults/.../internal/json/JsonUtils.java entirely:

- rowToJson: repoint all callers (DefaultJsonHandler, Flink DeltaWriterResult/
  DeltaCommittable/AbstractKernelTable, Spark v2 SerializableKernelRowWrapper,
  and their tests) to kernel-api io.delta.kernel.internal.util.JsonUtils.
- rowFromJson: move to a new static factory DefaultJsonRow.fromJsonString.
  It can't live in kernel-api because it constructs DefaultJsonRow (a
  kernel-defaults type) and parses with an unshaded Jackson ObjectMapper,
  whereas kernel-api shades Jackson. Behavior is unchanged from the old
  rowFromJson (same ObjectMapper.readTree -> new DefaultJsonRow, same
  RuntimeException on parse failure).
- Rename JsonUtilsSuite -> DefaultJsonRowSuite under internal/data to match
  the new home; the round-trip assertions are unchanged.

Validated (JDK17): kernel-defaults/flink/sparkV2 compile + checkstyle +
scalastyle + scalafmt clean; DefaultJsonRowSuite 12/12, RowTrackingUtilsTest
10/10, DeltaWriterResultTest + DeltaCommittableTest 2/2.
