+++ flink/src/main/java/io/delta/flink/sink/DeltaCommittable.java
@@ -17,8 +17,9 @@
 package io.delta.flink.sink;
 
 import io.delta.kernel.data.Row;
-import io.delta.kernel.defaults.internal.json.JsonUtils;
+import io.delta.kernel.defaults.internal.data.DefaultJsonRow;
 import io.delta.kernel.internal.actions.SingleAction;
+import io.delta.kernel.internal.util.JsonUtils;
 import io.delta.kernel.internal.util.Preconditions;
 import java.io.*;
 import java.util.ArrayList;
@@ -149,7 +150,7 @@ public class DeltaCommittable {
         List<Row> actions = new ArrayList<>(numActions);
         for (int i = 0; i < numActions; i++) {
           final String actionJson = in.readUTF();
-          actions.add(JsonUtils.rowFromJson(actionJson, SingleAction.FULL_SCHEMA));
+          actions.add(DefaultJsonRow.fromJsonString(actionJson, SingleAction.FULL_SCHEMA));
         }
         return new DeltaCommittable(jobId, operatorId, checkpointId, actions, context);
       } catch (ClassNotFoundException e) {
